+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/service/LoanStatusChangePlatformServiceImpl.java
@@ -25,14 +25,13 @@ import org.apache.fineract.infrastructure.event.business.BusinessEventListener;
 import org.apache.fineract.infrastructure.event.business.domain.loan.LoanStatusChangedBusinessEvent;
 import org.apache.fineract.infrastructure.event.business.service.BusinessEventNotifierService;
 import org.apache.fineract.portfolio.loanaccount.domain.Loan;
-import org.apache.fineract.portfolio.loanaccount.domain.LoanAccountDomainServiceJpa;
 
 @Slf4j
 @RequiredArgsConstructor
 public class LoanStatusChangePlatformServiceImpl implements LoanStatusChangePlatformService {
 
     private final BusinessEventNotifierService businessEventNotifierService;
-    private final LoanAccountDomainServiceJpa loanAccountDomainService;
+    private final LoanAccrualsProcessingService loanAccrualsProcessingService;
 
     @PostConstruct
     public void addListeners() {
@@ -47,7 +46,7 @@ public class LoanStatusChangePlatformServiceImpl implements LoanStatusChangePlat
             log.debug("Loan Status change for loan {}", loan.getId());
             if (loan.getStatus().isClosedObligationsMet() || loan.getStatus().isOverpaid()) {
                 log.debug("Loan Status {} for loan {}", loan.getStatus().getCode(), loan.getId());
-                loanAccountDomainService.applyFinalIncomeAccrualTransaction(loan);
+                loanAccrualsProcessingService.processAccrualsForLoanClosure(loan);
             }
             if (loan.isOpen()) {
                 loan.handleMaturityDateActivate();
